Updating CGNS and HDF5 versions.
Updating to HDF5 1.8.9 (since they fixed the cmake install rule that was broken
once upon a time, we don't need to apply any patches).

Updating to CGNS 3.1. Updating the patch for CGNS CMakeLists to set
target_link_libraries correctly when HDF5 support is enabled. This makes it
possible for us to build CGNS with HDF5 support.