Fix build errors when building numpy due to libpython*.so missing.
We need to pass LD_LIBRARY_PATH when launching python so it can find the
libpython*.so. However, with Ubuntu (and systems where cmake is build with
shared libs), that can cause problems since cmake ends up finding the libz we
built as part of our super-build.  Adding an advanced option
PASS_LD_LIBRARY_PATH_FOR_BUILDS that can be used to turn off the passing of
LD_LIBRARY_PATH on such systems.

Change-Id: I79861160fb9f92d975f85c4ec0537a79243b0e5f