All Build_Commands must be executed through external process.
when a user was specifiying a custom build_command we removed it
causing the default make command to happen. The correct solution is
that the custom build step needs to be called from the external process.
The only time this shouldn't happen is when the BUILD_COMMAND is equal
to an empty string which signifies that no build step should happen.

To further complicate the problem, the appending of the argument list
was causing collapsing of trailing empty list items i.e command;; which
made empty build_commands impossible. The fix is to use the
list(APPEND ...) call which doesn't remove empty list elements.

Change-Id: If6093617443114293e6e48247de4da28c3c1901a