Working on fixing MPI on Windows.
Updating to newer version of OpenMPI. Also figured out the bug in OpenMPI cmake
files that was resulting in failed ninja builds (and others too). Added a patch
for that.

Also static OpenMPI with shared ParaView fails. So we need to build shared
OpenMPI which requires compiler flags passed on to ParaView. To enable that,
added a mechanism to propagate compiler flags to dependent projects.

Change-Id: Ia2318b7e863f73b615e554554c701bf63040167a