When OPTIONAL_DEPENDS were specified, ensure that module order is correct.
We need to ensure that the processing order for modules is correct even after
optional modules have been enabled. This overcomes random cmake configuration
errors when modules were not enabled in the right order.

Also added a dependency on boost for paraview. The dummy module currently
discards all dependencies and hence we never ended up adding dependency
for boost on paraview. This overcomes that issue.

Change-Id: I7172acd10a38f798ddb17d246107e55dcc2cc3a2