Fix mpich issue when loading CosmologyTools plugin
Added some symbolic links to the bundled mpich libraries
to resolve issues when loading the CosmologyTools plugins.
When CosmologyTools plugins were loaded an error was thrown
about not being able to find libpmpich.dylib and libmpicxx.dylib.
Creating these symbolic links is a temporary workaround, but, likely not a good
long-term solution.

Change-Id: I185780fa0f2aad70394c8aa3f92ab92ddeef7f41