Fixes build issue on Windows with matplotlib.
On Windows nightly builds, especially with Ninja generator, we'd get errors of
the following sort:

----------------------------------------
...
[76/155] No update step for 'matplotlib'
[77/155] Performing patch step for 'matplotlib'
Error copying file (if different) from "C:/DBD/pvs-x64/install/lib/libpng14.lib"
                                        to "C:/DBD/pvs-x64/install/lib/png.lib".
[78/155] No update step for 'png'
[79/155] Performing configure step for 'png'
...
----------------------------------------

This was primarily because the patch-step for matplotlib was being executed
before the build/install-step for png. This was despite the fact that we had
added a dependency on "png" project for matplotlib. On inspection of the
generated build.ninja file, we realize that download/patch steps of any
external-project don't depend on the other projects that the project depends on.
It's only at configure-step that there are checks to ensure all dependended
projects are done. Thus accessing "png/zlib" files in patch step was a mistake.
Fixed by moving the patch step to be a configure step instead. This ensure all
dependencies are built before we attempt to move the files around.

Change-Id: I3d1c85c4e179d3bb03789488e8467693d1680ac4