Fixup matplotlib on apple.
This patch installs the png libraries into the bundle and tells
matplotlib how to find them at runtime. A new CMake variable is added,
PV_PYTHON_LIB_INSTALL_PREFIX, which refers to the location of the
python libraries in the bundle.

Note that this along doesn't fix matplotlib in mac bundles -- the
MPLFix branch of ParaView is also needed.

Change-Id: If59c41760741704437fe340b248890e0cf6fb754