Modifying all plugins to have versioned suffix + directory
--------------------------------------------------
Moving pv version info to a seperate cmake file.
--------------------------------------------------
The version information is updated with each new PV release. It is best
if this was in a seperate file.

We add another variable called pv_version_long which is the long suffix
version. This varialble will be used to name directories tarballs of
various plugins.

--------------------------------------------------
Including paraview_version in the top level.
--------------------------------------------------
This pv_version_information is make accessible at the top level. This
will enable various plugin to name their tarballs using this information.

--------------------------------------------------
Modifying all plugins to have versioned suffix + directory
--------------------------------------------------
There is no mechanism to differntiate plugin builds to the ParaView
version they build against. This is potentially problematic when
creating new tarballs for latest releases with the potential for the
older plugins being brought in by accident.

To solve this issue, plugin's tarballs should be genereated with a
proper suffix corriesponding to paraview version it is build against.
We thus build and copy the plugin into a directiory with appropriate
name and suffix. The directory is then zipped into a tarball with the
same name and suffix.

Change-Id: Ice4020bf1ada1b9ee1292bdd8daffafc5644b77d