Adding the option of downloading PV from another tarball source
Currently ParaView is downloaded from git repo at paraview.org. However
one may want to use another tarball from another source or even one in
the local filesystem.

This commit gives the user such an option.

If the MD5 sum is specified then the downloaded tarball is verified with it.

Change-Id: I417cc4c56d24f722c61ff2d7772c3c0edb562083