Adding option to automatically build the Nektar reader plugin.
The Nektar reader requires BLAS and LAPACK so I need to build
those as well. I'm removing the external Nektar plugin as well
to avoid confusion. The nektar reader plugin requires a
fortran compiler so we need to make sure one exists before
we allow the nektar reader to be enabled.

Change-Id: I939c6e69b0850e84faf0e124d5ace3ca1768ef25