Updated superbuild to build native cosmotools
This commit adds the GenericIO project in to the superbuild and
modifies the paraview and cosmotools projects to reflect recent
changes in paraview that natively integrate the CosmoTools plugins.
Instead of building a separate plugin, the code now lives in the
ParaView source tree (under ParaViewCore/VTKExtensions/CosmoTools)
and builds by linking to GenericIO(libGenericIO.a) and CosmoTools(libcosmotools.a)
externally.

Change-Id: I96aa712f367a47eb4c34b9c1a10ce86f6614dfe6