Add IBM BlueGeneQ cross compilation target.
This commit adds two targets, a non working XLC compiler option and a
working GNU compiler option. The XLC compiler option fails because we
run out of linker Table Of Contents (TOC) space, I'm adding this config
because with other compiler revisions the -qminimaltoc flag may
do what it is supposed to do.

Commit also splits out configuration options for Python so that it
can be customized per target.

Developed on mira@ANL.

Change-Id: I7046f518e39ab4a1bcedbbdd4db352e8827ee0e4