Added optional test of a remotely installed paraviewweb instance.
Added a python script which is capable of downloading, unpacking,
and configuring the necessary binaries to run paraviewweb.

The new test is a standalone python script which uses selenium
(needs to be installed in the system python) to interact with
remote paraviewweb.  For now, if some dependency of this test
is not met, the test passes.  However, in the future we can use
CMake's new SKIP_RETURN_CODE feature instead of just allowing
the test to pass.

Change-Id: I650569d5ec99a672c8dec96b3dae37ee5ba875e4