ffmpeg: pass flags down to ffmpeg's build
Without this, ffmpeg links against AVFoundation which is not available
on Snow Leopard. Unfortunately, FFmpeg ignores CFLAGS and LDFLAGS
environment variables, so they need to be passed as arguments to
configure. This has some fun artifacts from the ExternalProject
mechanism where spaces in command arguments are near impossible to get
through properly. To get around this, quote the command to avoid CMake's
list parsing rules.

Change-Id: I56087fd801d37f964bb8c3cc5d37a09d34e6bb9b