Fix issue with Python and rpath.
The python executable built by the superbuild needs to have rpath set to
properly load the shared library when running the executable in
build/configure stages.

To support that, python.cmake uses the newly extended append_flags()
mechanism to pass flags to the compiler/linker to add the appropriate
rpath.

Change-Id: I64068bf8bd118f1f0c805c48fa9e5e390b525f86