matplotlib: set PKG_CONFIG_PATH to find our freetype
A system-installed freetype will get in the way of the superbuild's
freetype because matplotlib uses pkg-config to find it and without this,
it gets its information from the system.

This is required because distributions bring in freetype's development
files from packages such as Qt which is too onerous to ask not be
installed in the general case.

Change-Id: I99cf542bd08f7e0be1774f86a1abd175fda8b011