qt: work around a bug in Qt's build system
See the comment in the code for what the bug is and the dates there show
that it has been around for a long time. Since no fix has come from
neither upstream nor anyone else who has been hit with this, just add a
hack to get around the problem.

Change-Id: Ibee599369b76ecb7f92448e07b6e785dcbd0bfab