qt4: install and build in one step
Qt4 is failing to build on debian4dash because QtCLucene is not found
during the build in the right place (the install directory). It seems
that installing makes things work.

This probably succeeded in the past because CMake was passing `-i` down
and the build errors got ignored, but later worked during the install
step.