Patch netcdf to inform VTK about changes to the netcdf API.
VTK's third party netcdf is not consistent with upstream, as some
methods take 'size_t*' instead of 'long*' parameters.

Previously we worked around this by checking if system netcdf was being
used, but the ParaView superbuild installs a netcdf with the same
modifications as the 'system' version.

This patch adds a new define to the modified headers, so we'll know when
we're using the modified version of the library.