Don't install dependencies outside of dependency_root
When using the superbuild to install on a system instead of creating a
package, you generally don't want the libraries specified by
USE_SYSTEM_foo to get pacakged up into the install, since they will
remain on the system being installed on anyways.  This leaves the
current behavior as default but allows you to turn off the bundling of
external dependencies.