paraview: fix default version detection
Use the right default version based on the selection made. For git
clones, use the latest version number because we can only guess at what
it will be. For source directories, introspection will give us an exact
answer in case it is a git repo and a best-guess via `version.txt`
otherwise. Other source selections are version numbers and should be
used as-is.