Add contents to qt.conf file for macOS
The default Qt plugins directory is expected to be
'paraview.app/Contents/plugins', but we are packaging Qt plugins into
'paraview.app/Contents/Plugins'. On case-sensitive file systems on
macOS, the change in case of 'Plugins' causes the
platforms/libqcocoa.dylib plugin to not be found and leads to a crash.
Fix this by specifying our non-default Plugins directory in qt.conf.

Fixes #52