changes to fix ospray under cray toolchain
bump embree to latest release to get most of the fixes.
Apply a patch corresponding to embree commit 7c0f6fc for another.
With this it should work.

Now update and reapply the findtbb patch to apply to newer embree.
The patch is really copying over ospray's findtbb, which we modify
to work with pv superbuild's install layout.