mili: add Mili 15.1 project
This project downloads the Mili 15.1 library source code, extracts it,
applies a few patches needed to the configuration and source code, and
builds it. The build is a little unusual in that the Makefile
generated by the configure step is within a subdirectory. Hence, the
build and install steps are added by separate calls to
superbuild_project_add_step. Furthermore, the install command is
broken in the project's Makefile, so we install libraries and headers
manually instead.