vrpn: update to enable HID on Mac and Windows
Point to a newer source archive to support the HID interface. HID
enables vrpn_server to interface with various USB devices. The
vrpn-wait3-removal.patch is no longer needed with the updated code, so
it has been removed.

Support for linux adds a dependency on libusb. Configuring devices USB
devices on linux requires some tricky configuraiton as well, so we
skip enabling HID on linux for now.