backport a fix for a crash with module_mpi
VTK creates new shared OSPRay buffers and then immediately commits
them. This exposes a bug in module_mpi. This patch will appear in
module_mpi with OSPRay 2.6 or so and should be removed after that
appears in ParaView.

Note: the buffer usage pattern is also inefficient with OSPRay's
distributed mode rendering. VTK should be changes to avoid it
at some point to.