move super-build forward
wslink update to 1.2.0
---
Fix restrict keyword compile error

Fixes an issue with older compilers and restrict keyword.
See https://github.com/util-linux/util-linux/issues/1405.
This should be resolved in 2.37.3 and 2.38 when it is released and this
patch can be removed.
---
Update mpi.cmake to utilize UCX if requested

Changes are based on work from Kevin Dean (kevinedwarddean@gmail.com)
in an alternate MR that because outdated.