Support packaging python-wrapped vtk modules in plugins
This is a temporary fix to allow otherwise unsupported behavior.  By
default the CAVE plugin has been disabled in the superbuild because
the workaround does not work on mac os.  Because the workaround does
work on linux and mac, the CAVE plugin has been enabled in those
environments so that binaries continue to have the functionality.