superbuild: bump for default component fixes
The default component was set too late in the packaging project leaving
one instance with an "Unspecified" component. The WIX CPack generator
uses a directory-per-component when multiple exist which caused the path
limit to be hit.

Also brings in a fix for Windows when building OpenPMD which uses
`find_package(Python)`. Not observed in CI, just on a development
machine.

* Ben Boeckel: python3-FindPython-variables (paraview/common-superbuild!599)
* Ben Boeckel: default-bundle-fix (paraview/common-superbuild!600)